+++ app/src/main/java/com/codesoom/assignment/controllers/product/ProductController.java
@@ -2,10 +2,10 @@
 // /products -> Create, Read
 // /products/{id} -> Read, Update, Delete
 
-package com.codesoom.assignment.controllers;
+package com.codesoom.assignment.controllers.product;
 
-import com.codesoom.assignment.application.AuthenticationService;
-import com.codesoom.assignment.application.ProductService;
+import com.codesoom.assignment.application.authentication.AuthenticationService;
+import com.codesoom.assignment.application.product.ProductService;
 import com.codesoom.assignment.domain.Product;
 import com.codesoom.assignment.dto.ProductData;
 import org.springframework.http.HttpStatus;
